Added rte-editor.ts
