+++ core/tests/playwright-acceptance-tests/utilities/user/logged-out-user.ts
@@ -338,23 +338,6 @@ export class LoggedOutUser extends BaseUser {
   }
 
   /**
-   * Clicks an element using JavaScript's native click() method.
-   * This ensures Angular properly handles the event in its change detection
-   * cycle, which is more reliable than Puppeteer's simulated clicks for
-   * Angular components like the sidebar.
-   * @param {string} selector - The CSS selector of the element to click.
-   */
-  private async clickWithJavaScript(selector: string): Promise<void> {
-    await this.waitForElementToStabilize(selector);
-    await this.page.evaluate((sel: string) => {
-      const element = document.querySelector(sel) as HTMLElement;
-      if (element) {
-        element.click();
-      }
-    }, selector);
-  }
-
-  /**
    * Function to close the hint modal.
    */
   async closeHintModal(): Promise<void> {
